# contract/http/openapi.yaml
openapi: 3.1.0
info:
  title: Tool-Kit Conversion Service
  version: 0.5.0
  description: >-
    CPU-only contract for durable, authenticated document conversion.
    Accepted conversions, idempotency records, and published artifacts are
    stored on disk and survive a service restart. Loopback-only.
servers:
  - url: http://127.0.0.1:8080
    description: Loopback-only development
paths:
  /health/live:
    get:
      operationId: getLiveness
      summary: Report whether the HTTP process is alive
      responses:
        "200":
          $ref: "#/components/responses/Health"
  /health/ready:
    get:
      operationId: getReadiness
      summary: Report whether the database, data root, and worker are usable
      responses:
        "200":
          description: Every dependency check passed
          headers:
            X-Request-Id:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReadinessResponse"
        "503":
          description: At least one dependency check failed
          headers:
            X-Request-Id:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReadinessResponse"
  /api/v1/capabilities:
    get:
      operationId: getCapabilities
      summary: Report only capabilities currently implemented
      responses:
        "200":
          description: Current service capabilities
          headers:
            X-Request-Id:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CapabilitiesEnvelope"
  /api/v1/conversions:
    post:
      operationId: createConversion
      summary: Stage a document and create a durable asynchronous conversion
      security:
        - bootstrapBearer: []
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/ConversionSubmission"
            encoding:
              source:
                contentType: application/pdf, application/vnd.openxmlformats-officedocument.wordprocessingml.document, application/msword, application/vnd.openxmlformats-officedocument.presentationml.presentation, application/vnd.ms-powerpoint, application/vnd.openxmlformats-officedocument.spreadsheetml.sheet, application/vnd.ms-excel, application/epub+zip, application/vnd.oasis.opendocument.text, application/vnd.oasis.opendocument.spreadsheet, application/vnd.oasis.opendocument.presentation, application/rtf, text/csv, application/vnd.ms-word.document.macroEnabled.12, application/vnd.ms-excel.sheet.macroEnabled.12, application/vnd.ms-powerpoint.presentation.macroEnabled.12, application/vnd.openxmlformats-officedocument.presentationml.slideshow, application/vnd.ms-powerpoint.slideshow.macroEnabled.12, image/png, image/jpeg, image/webp, image/tiff, image/gif, image/bmp, audio/wav, audio/mp4, video/mp4, video/quicktime, audio/mpeg, audio/flac
                headers:
                  Content-Disposition:
                    required: true
                    description: >-
                      The source part must carry a filename. The service reads
                      the format from the filename extension, so a part without
                      one is rejected with 422 `missing_filename`.


                      The extension must also be the one registered for the
                      `Content-Type` sent with this part. An extension that
                      names no supported format is 415
                      `unsupported_source_extension`, and an extension that
                      disagrees with the type is 415
                      `invalid_source_media_type`. Requiring both is what stops
                      a mislabeled upload from reaching an engine that cannot
                      read it.
                    schema:
                      type: string
                    example: form-data; name="source"; filename="report.pdf"
      responses:
        "202":
          description: Upload completed and the conversion was accepted
          headers:
            X-Request-Id:
              $ref: "#/components/headers/RequestId"
            Location:
              $ref: "#/components/headers/Location"
            Retry-After:
              $ref: "#/components/headers/RetryAfter"
            Idempotency-Replayed:
              $ref: "#/components/headers/IdempotencyReplayed"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobEnvelope"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "408":
          $ref: "#/components/responses/RequestTimeout"
        "409":
          $ref: "#/components/responses/Conflict"
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/conversions/{id}:
    get:
      operationId: getConversion
      summary: Poll one durable conversion
      security:
        - bootstrapBearer: []
      parameters:
        - $ref: "#/components/parameters/ConversionId"
      responses:
        "200":
          description: Current conversion state
          headers:
            X-Request-Id:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/conversions/{id}/artifacts:
    get:
      operationId: listConversionArtifacts
      summary: List atomically published artifacts
      security:
        - bootstrapBearer: []
      parameters:
        - $ref: "#/components/parameters/ConversionId"
      responses:
        "200":
          description: Empty until publication; two entries after success
          headers:
            X-Request-Id:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArtifactListEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/conversions/{id}/artifacts/markdown:
    get:
      operationId: downloadConversionMarkdown
      summary: Stream the published Markdown artifact
      security:
        - bootstrapBearer: []
      parameters:
        - $ref: "#/components/parameters/ConversionId"
      responses:
        "200":
          description: Published Markdown
          headers:
            X-Request-Id:
              $ref: "#/components/headers/RequestId"
            ETag:
              $ref: "#/components/headers/ArtifactEtag"
          content:
            text/markdown:
              schema:
                type: string
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/conversions/{id}/artifacts/manifest:
    get:
      operationId: downloadConversionManifest
      summary: Stream the published provenance manifest
      security:
        - bootstrapBearer: []
      parameters:
        - $ref: "#/components/parameters/ConversionId"
      responses:
        "200":
          description: Versioned provenance manifest
          headers:
            X-Request-Id:
              $ref: "#/components/headers/RequestId"
            ETag:
              $ref: "#/components/headers/ArtifactEtag"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConversionManifest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalError"
components:
  securitySchemes:
    bootstrapBearer:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: M1 bootstrap token loaded from a mounted secret file
  parameters:
    ConversionId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: Durable per-token record; 1-128 URL-safe characters
      schema:
        type: string
        minLength: 1
        maxLength: 128
        pattern: "^[A-Za-z0-9._~-]+$"
  headers:
    RequestId:
      description: Server-generated correlation identifier
      required: true
      schema:
        type: string
        format: uuid
    Location:
      description: Polling URL for the accepted conversion
      required: true
      schema:
        type: string
    RetryAfter:
      description: Suggested polling delay in seconds
      required: true
      schema:
        type: integer
        const: 1
    IdempotencyReplayed:
      description: Whether this response reused a stored idempotency record
      required: true
      schema:
        type: boolean
    ArtifactEtag:
      description: Quoted SHA-256-backed artifact entity tag
      required: true
      schema:
        type: string
  responses:
    Health:
      description: Service health
      headers:
        X-Request-Id:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/HealthResponse"
    Unauthorized:
      description: Missing or invalid bearer token
      headers:
        X-Request-Id:
          $ref: "#/components/headers/RequestId"
        WWW-Authenticate:
          schema:
            type: string
            const: Bearer
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    RequestTimeout:
      description: Upload deadline exceeded
      headers:
        X-Request-Id:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    Conflict:
      description: >-
        Idempotency, profile, engine availability, or artifact-state conflict.
        `source_format_unavailable` means the format is in the contract and no
        engine on this deployment converts it; check `inputFormats` first.
      headers:
        X-Request-Id:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    PayloadTooLarge:
      description: Upload exceeded configured byte limits
      headers:
        X-Request-Id:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    UnsupportedMediaType:
      description: Source extension, declared media type, or content signature failed
      headers:
        X-Request-Id:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    UnprocessableEntity:
      description: Multipart field semantics are invalid
      headers:
        X-Request-Id:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    TooManyRequests:
      description: Active job capacity reached
      headers:
        X-Request-Id:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    NotFound:
      description: Conversion, artifact, or route not found
      headers:
        X-Request-Id:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    InternalError:
      description: Safe internal failure without parser details
      headers:
        X-Request-Id:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    BadRequest:
      description: Malformed request
      headers:
        X-Request-Id:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
  schemas:
    HealthResponse:
      type: object
      required: [status, service, serviceVersion]
      properties:
        status:
          type: string
        service:
          type: string
          const: tool-kit-converter
        serviceVersion:
          type: string
    ReadinessResponse:
      type: object
      required: [status, service, serviceVersion, checks]
      properties:
        status:
          type: string
          enum: [ready, not_ready]
        service:
          type: string
          const: tool-kit-converter
        serviceVersion:
          type: string
        checks:
          type: object
          required: [database, dataRoot, worker]
          additionalProperties: false
          properties:
            database:
              type: string
              enum: [ok, failed]
            dataRoot:
              type: string
              enum: [ok, failed]
            worker:
              type: string
              enum: [ok, failed]
    CapabilitiesEnvelope:
      type: object
      required: [data]
      properties:
        data:
          type: object
          required: [apiVersion, serviceVersion, conversion, remoteFallback]
          properties:
            apiVersion:
              type: string
              const: v1
            serviceVersion:
              type: string
            conversion:
              $ref: "#/components/schemas/ConversionCapabilities"
            remoteFallback:
              type: object
              required: [provider, available]
              properties:
                provider:
                  type: string
                  const: datalab
                available:
                  type: boolean
                  const: false
    ConversionCapabilities:
      type: object
      required: [acceptingJobs, durability, inputFormats, outputFormats, profiles, engines, limits]
      properties:
        acceptingJobs:
          type: boolean
        durability:
          type: string
          const: persistent
        inputFormats:
          description: >-
            What this deployment can convert, not what the contract knows. The
            image types need the Apple Vision engine and the audio types the
            Apple speech engine, both of which ship only with macOS 26 and
            later, so a deployment without them advertises the other 18 and
            refuses those uploads with 409 `source_format_unavailable`. Read
            this list rather than assuming the whole enum.


            Table order: the 18 document types first, then the image types,
            then the audio and video ones.
          type: array
          items:
            type: string
            enum:
              - application/pdf
              - application/vnd.openxmlformats-officedocument.wordprocessingml.document
              - application/msword
              - application/vnd.openxmlformats-officedocument.presentationml.presentation
              - application/vnd.ms-powerpoint
              - application/vnd.openxmlformats-officedocument.spreadsheetml.sheet
              - application/vnd.ms-excel
              - application/epub+zip
              - application/vnd.oasis.opendocument.text
              - application/vnd.oasis.opendocument.spreadsheet
              - application/vnd.oasis.opendocument.presentation
              - application/rtf
              - text/csv
              - application/vnd.ms-word.document.macroEnabled.12
              - application/vnd.ms-excel.sheet.macroEnabled.12
              - application/vnd.ms-powerpoint.presentation.macroEnabled.12
              - application/vnd.openxmlformats-officedocument.presentationml.slideshow
              - application/vnd.ms-powerpoint.slideshow.macroEnabled.12
              - image/png
              - image/jpeg
              - image/webp
              - image/tiff
              - image/gif
              - image/bmp
              - audio/wav
              - audio/mp4
              - video/mp4
              - video/quicktime
              - audio/mpeg
              - audio/flac
          minItems: 18
          maxItems: 30
        outputFormats:
          type: array
          items:
            type: string
        profiles:
          type: array
          items:
            type: object
            required: [name, available]
            properties:
              name:
                $ref: "#/components/schemas/ConversionProfile"
              available:
                type: boolean
        engines:
          description: >-
            The engines running in this process, pdf-inspector then anydoc.
            apple-vision and local-audio follow, each present only where it
            initialized. apple-vision's version is the host's macOS product
            version, because Vision ships with the OS and has no version of its
            own; local-audio's is that version plus the speaker model package
            it was built against, `<macOS>+fluidaudio-<version>`.
          type: array
          items:
            type: object
            required: [name, version]
            properties:
              name:
                type: string
                enum: [pdf-inspector, anydoc, apple-vision, local-audio]
              version:
                type: string
          minItems: 2
          maxItems: 4
        limits:
          type: object
          required:
            [
              maxUploadBytes,
              maxAudioUploadBytes,
              maxOutputBytes,
              maxActiveJobs,
              maxConcurrentUploads,
            ]
          properties:
            maxUploadBytes:
              description: >-
                The ceiling for every source except audio, which is bounded by
                maxAudioUploadBytes.
              type: integer
              minimum: 1
            maxAudioUploadBytes:
              type: integer
              minimum: 1
            maxOutputBytes:
              type: integer
              minimum: 1
            maxActiveJobs:
              type: integer
              minimum: 1
            maxConcurrentUploads:
              type: integer
              minimum: 1
    ConversionSubmission:
      type: object
      required: [source, clientRunId, profile]
      additionalProperties: false
      properties:
        source:
          type: string
          format: binary
          description: >-
            The document bytes. The part must be non-empty and must open with
            the container signature of the format its extension names, or the
            service answers 415 `invalid_source_signature`. The signature is
            checked against the bytes actually received, so a renamed file is
            rejected at admission rather than at conversion.


            Two pieces of metadata travel with the bytes and are required as
            well, a filename and a `Content-Type`. Both are described under
            `encoding.source`.
        clientRunId:
          type: string
          format: uuid
        profile:
          $ref: "#/components/schemas/ConversionProfile"
        languageCorrection:
          type: boolean
          description: >-
            Whether local OCR applies language correction to the text it
            recognizes. Omit the part to leave it on. A value other than
            `true` or `false` is 422 `invalid_language_correction`.


            Read only by the local image OCR engine and ignored by every other
            route, so sending it with a PDF or an Office document changes
            nothing.
        customWords:
          type: string
          description: >-
            Words local OCR should prefer when it is unsure, one per line.
            Omit the part to send none. The field shares the 256-byte cap
            every text part has, and exceeding it is 422 `metadata_too_large`.


            Read only by the local image OCR engine and ignored by every other
            route.
        speakerCount:
          type: integer
          minimum: 1
          maximum: 20
          description: >-
            How many people speak in the recording. Omit the part and the
            diarizer guesses, which the transcript says in its first line. A
            value outside 1 to 20, or one that is not an integer, is 422
            `invalid_speaker_count`.


            Read only by the local audio engine and ignored by every other
            route.
    ConversionProfile:
      type: string
      enum: [standard, local_only, best_quality]
    JobEnvelope:
      type: object
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/ConversionJob"
    ConversionJob:
      type: object
      required:
        - id
        - activeAttemptId
        - clientRunId
        - profile
        - status
        - warnings
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
        activeAttemptId:
          type: string
          format: uuid
        clientRunId:
          type: string
          format: uuid
        profile:
          $ref: "#/components/schemas/ConversionProfile"
        status:
          type: string
          enum: [queued, converting_local, finalizing, succeeded, failed, needs_remote]
        route:
          $ref: "#/components/schemas/Route"
        warnings:
          type: array
          items:
            $ref: "#/components/schemas/Warning"
        failure:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ArtifactListEnvelope:
      type: object
      required: [data]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Artifact"
    Artifact:
      type: object
      required: [kind, attemptId, mediaType, byteLength, sha256, href]
      properties:
        kind:
          type: string
          enum: [markdown, manifest]
        attemptId:
          type: string
          format: uuid
        mediaType:
          type: string
        byteLength:
          type: integer
          minimum: 1
        sha256:
          $ref: "#/components/schemas/Sha256"
        href:
          type: string
    ConversionManifest:
      type: object
      required:
        - schemaVersion
        - jobId
        - attemptId
        - clientRunId
        - profile
        - source
        - engine
        - route
        - document
        - warnings
        - output
        - startedAt
        - completedAt
      properties:
        schemaVersion:
          type: integer
          const: 1
        jobId:
          type: string
          format: uuid
        attemptId:
          type: string
          format: uuid
        clientRunId:
          type: string
          format: uuid
        profile:
          $ref: "#/components/schemas/ConversionProfile"
        source:
          $ref: "#/components/schemas/ManifestArtifact"
        engine:
          type: object
          required: [name, version, features]
          properties:
            name:
              type: string
              enum: [pdf-inspector, anydoc, apple-vision, local-audio]
            version:
              type: string
            features:
              type: array
              maxItems: 0
              items:
                type: string
        route:
          $ref: "#/components/schemas/Route"
        document:
          description: >-
            Engine diagnostics. Inspection when engine.name is pdf-inspector,
            AnyDocDiagnostics when it is anydoc, VisionDiagnostics when it is
            apple-vision, AudioDiagnostics when it is local-audio.
          oneOf:
            - $ref: "#/components/schemas/Inspection"
            - $ref: "#/components/schemas/AnyDocDiagnostics"
            - $ref: "#/components/schemas/VisionDiagnostics"
            - $ref: "#/components/schemas/AudioDiagnostics"
        warnings:
          type: array
          items:
            $ref: "#/components/schemas/Warning"
        output:
          $ref: "#/components/schemas/ManifestArtifact"
        startedAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
    ManifestArtifact:
      type: object
      required: [mediaType, byteLength, sha256]
      properties:
        mediaType:
          type: string
        byteLength:
          type: integer
          minimum: 1
        sha256:
          $ref: "#/components/schemas/Sha256"
    Inspection:
      type: object
      required:
        - pdfType
        - confidence
        - pageCount
        - pagesNeedingOcr
        - ocrReasonsByPage
        - hasEncodingIssues
        - isComplex
        - pagesWithTables
        - pagesWithColumns
        - processingTimeMs
      properties:
        pdfType:
          type: string
          const: text_based
        confidence:
          type: number
          minimum: 0
          maximum: 1
        pageCount:
          type: integer
          minimum: 1
        pagesNeedingOcr:
          type: array
          maxItems: 0
          items:
            type: integer
        ocrReasonsByPage:
          type: array
          items:
            type: object
        hasEncodingIssues:
          type: boolean
          const: false
        isComplex:
          type: boolean
        pagesWithTables:
          type: array
          items:
            type: integer
        pagesWithColumns:
          type: array
          items:
            type: integer
        processingTimeMs:
          type: integer
          minimum: 0
    Route:
      type: object
      required: [kind, reasonCodes]
      properties:
        kind:
          type: string
          enum: [local_pdf, local_anydoc, local_vision, local_audio]
        reasonCodes:
          description: >-
            Why the conversion routed the way it did, in decision order.
            native_text_pdf, structured_document, recognized_image_text, and
            transcribed_audio are the policy's own codes, one per engine that
            published. The rest are the reason an engine gave for producing no
            Markdown, which always ends the job needs_remote.


            Open vocabulary, deliberately not an enum: the remote route adds
            codes, and a closed enum would make that a breaking change for
            every generated client. Render an unknown code verbatim.
          type: array
          items:
            type: string
            x-vocabulary:
              - native_text_pdf
              - structured_document
              - recognized_image_text
              - transcribed_audio
              - scanned_pdf
              - image_based_pdf
              - mixed_pdf
              - garbled_text
              - ocr_required
              - local_quality_failed
              - output_too_large
    Warning:
      description: >-
        A caveat about output that was published anyway. Warnings never change
        the route, they exist so a degraded success stops being a silent one.
        pages_without_extractable_text means some pages produced no text; the
        engine cannot say whether those pages were blank or a scan whose
        content is now missing, so treat it as a prompt to check rather than as
        proof of loss. dense_tables and multi_column_layout mean local
        reconstruction of that layout is weaker than the remote route's.


        Open vocabulary on the same terms as Route.reasonCodes.
      type: string
      x-vocabulary:
        - pages_without_extractable_text
        - dense_tables
        - multi_column_layout
    AnyDocDiagnostics:
      type: object
      required: [format, processingTimeMs]
      properties:
        format:
          type: string
        processingTimeMs:
          type: integer
          minimum: 0
    VisionDiagnostics:
      type: object
      required: [processingTimeMs]
      # Closed, because wall time is the whole of it and an open object here
      # would also match the other three engines' diagnostics.
      additionalProperties: false
      properties:
        processingTimeMs:
          type: integer
          minimum: 0
    AudioDiagnostics:
      description: >-
        Content-free measurements of one transcription: durations, the speaker
        count the diarizer settled on and whether it guessed it, never words.
      type: object
      required:
        - processingTimeMs
        - audioSeconds
        - speakersFound
        - speakerCountGuessed
        - locale
      additionalProperties: false
      properties:
        processingTimeMs:
          type: integer
          minimum: 0
        audioSeconds:
          type: number
          minimum: 0
        speakersFound:
          type: integer
          minimum: 1
        speakerCountGuessed:
          type: boolean
        locale:
          type: string
    Sha256:
      type: string
      pattern: "^[a-f0-9]{64}$"
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, requestId, details]
          properties:
            code:
              type: string
            message:
              type: string
            requestId:
              type: string
              format: uuid
            details:
              type: array
              maxItems: 0
              items:
                type: object